+++ Classes/OldModel/LegacySpeaker.php
@@ -8,6 +8,8 @@ use OliverKlee\Oelib\Mapper\MapperRegistry;
 use OliverKlee\Seminars\Mapper\FrontEndUserMapper;
 use OliverKlee\Seminars\Model\FrontEndUser;
 use TYPO3\CMS\Core\Resource\FileReference;
+use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
+use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 
 /**
  * This class represents a speaker.
@@ -205,16 +207,15 @@ class LegacySpeaker extends AbstractModel
     public function getLinkedTitle(): string
     {
         $encodedTitle = \htmlspecialchars($this->getTitle(), ENT_QUOTES | ENT_HTML5);
-        if (!$this->hasHomepage()) {
-            return $encodedTitle;
+        $frontEndController = $GLOBALS['TSFE'] ?? null;
+        $contentObject = $frontEndController instanceof TypoScriptFrontendController ? $frontEndController->cObj : null;
+        if ($contentObject instanceof ContentObjectRenderer && $this->hasHomepage()) {
+            $result = $contentObject->getTypoLink($encodedTitle, $this->getHomepage());
+        } else {
+            $result = $encodedTitle;
         }
 
-        $encodedUrl = \htmlspecialchars(
-            $this->addMissingProtocolToUrl($this->getHomepage()),
-            ENT_QUOTES | ENT_HTML5
-        );
-
-        return '<a href="' . $encodedUrl . '">' . $encodedTitle . '</a>';
+        return $result;
     }
 
     /**